Fixed printWorld, added a pause before traversal
Fixed printing out the grid.
Added a scanner to take an input before continuing to explode the
console with traversal statements.

NOTE: QLearning does not work sometimes. In the test provided with this
commit, beginning at (0,0) and reward at (0,2) does not work, but
beginning at (0,0) and reward at (2,3) does.

Could swear that it didn’t work at reward (1,0) or (0,1) but it does
now. So maybe it’s just randomness. Either way, bad for a grid that
small.

Also traversal confirmed to be broken- the qtable being produced is
correct (when it produces SOMETHING, if it’s wrong, all values are 0).