Actually move states in traversal
Was not moving states in traversal, resulting in it picking the same
direction every time and staying at the start state.

The algorithm is not appropriately selecting the correct direction to
move in- it picks down, which has reward of 25. Right has a reward of
50. Therefore the max was not selected.

changed == to compareTo for sanity.
changed the learning parameters for sanity.