Bugfix: max reward not reset each step
Resetting the max reward each time the agent moves in traversal.

TODO:
Neighbor positions are wrong, see this output:

(1,0)
reward for right is 0.0
reward for left is 0.0
reward for down is 25.0
reward for up is 25.0
Agent moves in direction:down
Moved from (1,0) to (1,2)

How does one move from (1,0) to (1,2) in one step?

Also, the agent hangs here and continues to stay in this location.